[IPR-39] refactor(recruitment): Area enum 경로 이동
