+++ src/main/java/com/flex/interpre/domain/recruitment/dto/response/RecruitmentResponse.java
@@ -2,10 +2,7 @@ package com.flex.interpre.domain.recruitment.dto.response;
 
 import com.flex.interpre.domain.company.dto.CompanySummaryResponse;
 import com.flex.interpre.domain.recruitment.entity.*;
-import com.flex.interpre.global.constant.Area;
-import com.flex.interpre.global.constant.JobSecond;
-import com.flex.interpre.global.constant.JobFirst;
-import com.flex.interpre.global.constant.JobThird;
+import com.flex.interpre.global.constant.*;
 import lombok.Builder;
 
 import java.time.LocalDateTime;
