+++ src/main/java/com/flex/interpre/domain/recruitment/entity/Recruitment.java
@@ -2,10 +2,7 @@ package com.flex.interpre.domain.recruitment.entity;
 
 import com.flex.interpre.domain.recruitment.dto.request.RecruitmentCreateUpdateRequest;
 import com.flex.interpre.domain.company.entity.Company;
-import com.flex.interpre.global.constant.Area;
-import com.flex.interpre.global.constant.JobSecond;
-import com.flex.interpre.global.constant.JobFirst;
-import com.flex.interpre.global.constant.JobThird;
+import com.flex.interpre.global.constant.*;
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
